+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added restriction to require `-j` or `--json-export` flag when using pipe redirection
+- Added undo and redo functionality, supporting operation history for cell editing, row/column deletion, and sheet deletion
 
 ### Changed